+++ src/Libraries/SmartStore.Services/Security/AclService.cs
@@ -271,8 +271,8 @@ namespace SmartStore.Services.Security
 
         private string GetSegmentKeyPart(string entityName, int entityId, out int minId, out int maxId)
         {
-            maxId = entityId.GetRange(1000, out minId);
-            return (entityName + "." + maxId.ToString()).ToLowerInvariant();
+            (minId, maxId) = entityId.GetRange(1000);
+            return (entityName + "." + minId.ToString()).ToLowerInvariant();
         }
 
         #endregion